End of Section 10

+++ Application/Activities/Command/DeleteActivity.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Net;
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -6,18 +8,28 @@ namespace Application.Activities.Command;
 
 public class DeleteActivity
 {
-    public class Command : IRequest
+    public class Command : IRequest<Result<Unit>>
     {
         public required string Id { get; set; }
     }
 
-    public class Handler(AppDbContext context) : IRequestHandler<Command>
+    public class Handler(AppDbContext context) : IRequestHandler<Command, Result<Unit>>
     {
-        public async Task Handle(Command request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
         {
-            var activity = await context.Activities.FindAsync([request.Id], cancellationToken) ?? throw new Exception("Activity not found");
+            var activity = await context.Activities.FindAsync([request.Id], cancellationToken);
+            if (activity is null)
+            {
+                return Result<Unit>.Failure("Activity not found", (int)HttpStatusCode.NotFound);
+            }
             context.Activities.Remove(activity);
-            await context.SaveChangesAsync(cancellationToken);
+            var result = await context.SaveChangesAsync(cancellationToken);
+            if (result > 0)
+            {
+                return Result<Unit>.Success(Unit.Value);
+            }
+
+            return Result<Unit>.Failure("Failure to remove activity", (int)HttpStatusCode.BadRequest);
         }
     }
